refactor(dashboard): rename totalBalance internals to portfolioOverview
Align component names, feature exports, and i18n keys with the
already-renamed dashboard-portfolio-overview directory.
