+++ src/renderer/features/dashboard-portfolio-overview/ui/ChainDetailModal.tsx
@@ -3,21 +3,21 @@ import { Cell, Pie, PieChart, Tooltip } from 'recharts';
 
 import { type CurrencyItem } from '@/shared/api/price-provider';
 import { useI18n } from '@/shared/i18n';
-import { formatBalance, toAddress, toShortAddress } from '@/shared/lib/utils';
+import { formatBalance } from '@/shared/lib/utils';
 import { FootnoteText } from '@/shared/ui';
-import { AssetIcon, Identicon } from '@/shared/ui-entities';
+import { AssetIcon } from '@/shared/ui-entities';
 import { type Column, Modal, Table } from '@/shared/ui-kit';
 
 import { Price } from './Price';
-import { CHART_TOOLTIP_STYLE, FALLBACK_COLORS } from './chartConstants';
-import { type BreakdownRow, useHoldingBreakdown } from './useHoldingBreakdown';
-import { type Holding } from './useHoldings';
+import { CHART_TOOLTIP_STYLE, getAssetColor } from './chartConstants';
+import { type ChainAssetRow, useChainBreakdown } from './useChainBreakdown';
+import { type ChainHolding } from './useChainHoldings';
 
 type ChartEntry = {
   name: string;
   value: number;
   index: number;
-  row: BreakdownRow;
+  row: ChainAssetRow;
 };
 
 type TooltipPayloadItem = {
@@ -31,61 +31,56 @@ const ChartTooltip = ({ active, payload }: { active?: boolean; payload?: Tooltip
   if (!item) return null;
 
   const { row } = item.payload;
+  const { formatted, suffix } = formatBalance(row.rawAmount, row.precision);
 
   return (
     <div style={CHART_TOOLTIP_STYLE}>
-      <div style={{ fontWeight: 600 }}>{row.name || toShortAddress(row.address)}</div>
+      <div style={{ fontWeight: 600 }}>{row.symbol}</div>
+      <div>
+        {formatted}
+        {suffix} {row.symbol}
+      </div>
       <div>{row.sharePercent.toFixed(1)}%</div>
     </div>
   );
 };
 
-type EntryLike = { accountId: string; name: string; address: string };
-
 type Props = {
-  holding: Holding;
+  chainHolding: ChainHolding;
   accountIds: string[];
-  allEntries: EntryLike[];
   currency: CurrencyItem | null;
   onClose: () => void;
 };
 
-export const AssetDetailModal = ({ holding, accountIds, allEntries, currency, onClose }: Props) => {
+export const ChainDetailModal = ({ chainHolding, accountIds, currency, onClose }: Props) => {
   const { t } = useI18n();
-  const { rows } = useHoldingBreakdown(holding.priceId, accountIds, allEntries);
-  const addressCount = rows.length;
-  const { formatted, suffix } = formatBalance(holding.totalRaw, holding.precision);
+  const { rows } = useChainBreakdown(chainHolding.chainId, accountIds);
 
   const chartData = useMemo<ChartEntry[]>(
     () =>
-      rows
-        .map((row, i) => ({ name: row.name || toShortAddress(row.address), value: row.fiatValueNum, index: i, row }))
-        .filter((d) => d.value > 0),
+      rows.map((row, i) => ({ name: row.symbol, value: row.fiatValueNum, index: i, row })).filter((d) => d.value > 0),
     [rows],
   );
 
-  const columns: Column<BreakdownRow>[] = [
+  const columns: Column<ChainAssetRow>[] = [
     {
       key: 'name',
-      title: t('dashboard.totalBalance.assetDetail.address'),
+      title: t('dashboard.portfolioOverview.chainDetail.asset'),
       width: '35%',
       render: (_, item) => (
         <div className="flex items-center gap-2">
           <span
             className="h-2 w-2 shrink-0 rounded-full"
-            style={{ backgroundColor: FALLBACK_COLORS[item.colorIndex % FALLBACK_COLORS.length] }}
+            style={{ backgroundColor: getAssetColor(item.priceId, item.colorIndex) }}
           />
-          <Identicon address={toAddress(item.address)} />
-          <div className="min-w-0">
-            <FootnoteText className="truncate font-semibold">{item.name}</FootnoteText>
-            <FootnoteText className="text-text-tertiary">{toShortAddress(item.address)}</FootnoteText>
-          </div>
+          <AssetIcon asset={item} size={24} />
+          <FootnoteText className="truncate font-semibold">{item.symbol}</FootnoteText>
         </div>
       ),
     },
     {
       key: 'rawAmountNum',
-      title: t('dashboard.totalBalance.assetDetail.amount'),
+      title: t('dashboard.portfolioOverview.chainDetail.amount'),
       sortable: true,
       width: '25%',
       render: (_, item) => {
@@ -101,7 +96,7 @@ export const AssetDetailModal = ({ holding, accountIds, allEntries, currency, on
     },
     {
       key: 'fiatValueNum',
-      title: t('dashboard.totalBalance.assetDetail.value'),
+      title: t('dashboard.portfolioOverview.chainDetail.value'),
       sortable: true,
       width: '22%',
       render: (_, item) => (
@@ -112,7 +107,7 @@ export const AssetDetailModal = ({ holding, accountIds, allEntries, currency, on
     },
     {
       key: 'sharePercent',
-      title: t('dashboard.totalBalance.assetDetail.share'),
+      title: t('dashboard.portfolioOverview.chainDetail.share'),
       sortable: true,
       width: '18%',
       render: (_, item) => <FootnoteText className="tabular-nums">{item.sharePercent.toFixed(1)}%</FootnoteText>,
@@ -123,23 +118,21 @@ export const AssetDetailModal = ({ holding, accountIds, allEntries, currency, on
 
   return (
     <Modal isOpen size="lg" onToggle={(open) => !open && onClose()}>
-      <Modal.Title close>{t('dashboard.totalBalance.assetDetail.title', { symbol: holding.symbol })}</Modal.Title>
+      <Modal.Title close>
+        {t('dashboard.portfolioOverview.chainDetail.title', { chainName: chainHolding.chainName })}
+      </Modal.Title>
       <Modal.Content disableScroll>
         <div className="flex items-center gap-3 px-5 py-3">
-          <AssetIcon asset={holding} size={32} />
+          <img src={chainHolding.chainIcon} alt={chainHolding.chainName} width={32} height={32} className="shrink-0" />
           <div className="min-w-0 flex-1">
-            <FootnoteText className="font-bold">{holding.symbol}</FootnoteText>
+            <FootnoteText className="font-bold">{chainHolding.chainName}</FootnoteText>
             <FootnoteText className="text-text-tertiary">
-              {t('dashboard.totalBalance.assetDetail.addressCount', { count: addressCount })}
+              {t('dashboard.portfolioOverview.chainDetail.assetCount', { count: chainHolding.assetCount })}
             </FootnoteText>
           </div>
           <div className="shrink-0">
             <FootnoteText align="right" className="font-bold tabular-nums">
-              {formatted}
-              {suffix} {holding.symbol}
-            </FootnoteText>
-            <FootnoteText align="right" className="text-text-tertiary tabular-nums">
-              <Price amount={holding.fiatValue} currency={currency} />
+              <Price amount={chainHolding.fiatValue} currency={currency} />
             </FootnoteText>
           </div>
         </div>
@@ -158,7 +151,7 @@ export const AssetDetailModal = ({ holding, accountIds, allEntries, currency, on
                 animationDuration={400}
               >
                 {chartData.map((entry) => (
-                  <Cell key={entry.index} fill={FALLBACK_COLORS[entry.index % FALLBACK_COLORS.length]} />
+                  <Cell key={entry.index} fill={getAssetColor(entry.row.priceId, entry.index)} />
                 ))}
               </Pie>
               <Tooltip content={<ChartTooltip />} />
