+++ src/renderer/features/dashboard-portfolio-overview/ui/chartConstants.ts
